Extract elements into new module.
